# openapi/bundled.yaml
openapi: 3.1.0
info:
  title: Chat API
  version: 0.1.0
  description: Slack-like communication application API
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/attachments/presign:
    post:
      operationId: presignUpload
      summary: Get presigned upload URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignRequest'
      responses:
        '200':
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/attachments/{id}:
    get:
      operationId: getAttachment
      summary: Get attachment metadata
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Attachment metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/attachments/{id}/download:
    get:
      operationId: downloadAttachment
      summary: Get presigned download URL
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Presigned download URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                    format: uri
                required:
                  - downloadUrl
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/refresh:
    post:
      operationId: refresh
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                required:
                  - accessToken
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/channels/{channelId}:
    patch:
      operationId: updateChannel
      summary: Update channel
      security:
        - bearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelRequest'
      responses:
        '200':
          description: Channel updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/links/fetch-ogp:
    post:
      operationId: fetchOGP
      summary: Fetch OGP data for a URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchOGPRequest'
      responses:
        '200':
          description: OGP data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchOGPResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/workspaces/{id}/channels:
    get:
      operationId: listChannels
      summary: List channels in workspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createChannel
      summary: Create a new channel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - messageId
        - fileName
        - mimeType
        - sizeBytes
        - createdAt
...
    Channel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isPrivate:
          type: boolean
        channelType:
          type: string
          enum:
            - public
            - private
            - dm
            - group_dm
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
          description: 未読メッセージ数
          minimum: 0
        hasMention:
          type: boolean
          description: 未読メンションの有無
      required:
        - id
        - workspaceId
        - name
        - isPrivate
        - createdBy
        - createdAt
...
    CreateChannelRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        isPrivate:
          type: boolean
          default: false
      required:
        - name
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    FetchOGPRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
      required:
        - url
    FetchOGPResponse:
      type: object
      properties:
        ogpData:
          type: object
          properties:
            title:
              type: string
              nullable: true
            description:
              type: string
              nullable: true
            imageUrl:
              type: string
              nullable: true
            siteName:
              type: string
              nullable: true
            cardType:
              type: string
              nullable: true
      required:
        - ogpData
...
    PresignRequest:
      type: object
      properties:
        fileName:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
        channelId:
          type: string
          format: uuid
      required:
        - fileName
        - contentType
        - sizeBytes
        - channelId
    PresignResponse:
      type: object
      properties:
        uploadUrl:
          type: string
          format: uri
        attachmentId:
          type: string
          format: uuid
      required:
        - uploadUrl
        - attachmentId
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
...
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        isPrivate:
          type: boolean
...